+++ openslides_backend/action/actions/meeting/import_.py
@@ -40,7 +40,7 @@ from ...util.register import register_action
 from ...util.typing import ActionData, ActionResultElement, ActionResults
 from ..meeting_user.helper_mixin import MeetingUserHelperMixin
 from ..motion.update import EXTENSION_REFERENCE_IDS_PATTERN
-from ..user.user_mixin import LimitOfUserMixin, UsernameMixin
+from ..user.user_mixins import LimitOfUserMixin, UsernameMixin
 
 
 @register_action("meeting.import")
