+++ openslides_backend/action/actions/user/password_mixins.py
@@ -1,7 +1,33 @@
 from typing import Any, Callable, Dict
 
+from ....shared.exceptions import ActionException
 from ...action import Action
+from ...generics.update import UpdateAction
 from ...util.typing import ActionData
+
+
+class UserSetPasswordMixin(UpdateAction):
+    def update_instance(self, instance: Dict[str, Any]) -> Dict[str, Any]:
+        """
+        set hashed password and set default password if set_as_default is True.
+        """
+        user = self.datastore.get(
+            f"user/{instance['id']}", ["saml_id"], lock_result=False
+        )
+        if user.get("saml_id"):
+            raise ActionException(
+                f"user {user['saml_id']} is a Single Sign On user and has no local Openslides passwort."
+            )
+
+        password = instance.pop("password")
+        set_as_default = False
+        if "set_as_default" in instance:
+            set_as_default = instance.pop("set_as_default")
+        hashed_password = self.auth.hash(password)
+        instance["password"] = hashed_password
+        if set_as_default:
+            instance["default_password"] = password
+        return instance
 
 
 class PasswordMixin(Action):
