+++ openslides_backend/action/actions/user/user_mixins.py
@@ -5,14 +5,14 @@ from typing import Any, Dict, List, Optional
 from openslides_backend.shared.typing import HistoryInformation
 from openslides_backend.shared.util import ONE_ORGANIZATION_FQID
 
-from ....action.action import Action, original_instances
-from ....action.mixins.archived_meeting_check_mixin import CheckForArchivedMeetingMixin
 from ....presenter.search_users import SearchUsers
 from ....services.datastore.interface import DatastoreService
 from ....shared.exceptions import ActionException
 from ....shared.filters import FilterOperator
 from ....shared.patterns import FullQualifiedId, fqid_from_collection_and_id
 from ....shared.schema import decimal_schema, id_list_schema, optional_id_schema
+from ...action import Action, original_instances
+from ...mixins.archived_meeting_check_mixin import CheckForArchivedMeetingMixin
 from ...util.typing import ActionData
 from ..meeting_user.set_data import MeetingUserSetData
